requirements.txt
chroma-hnswlib==0.7.6
chromadb==0.5.12
cohere==5.11.0
django-cors-headers==3.7.0
docker==7.1.0
groq==0.11.0
jsonschema==4.23.0
langchain==0.3.2
langchain-cohere==0.3.0
langchain-community==0.3.1
langchain-core==0.3.9
langchain-experimental==0.3.2
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.2
langsmith==0.1.132
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
uvicorn==0.31.0
openai==1.58.1
dnspython==2.6.1
pymongo==3.11.4
gunicorn==23.0.0
rapidfuzz==3.6.1
nltk==3.8.1
jieba==0.42.1
sentence-transformers==3.4.1
djongo5==1.3.9
django-filter==25.1
djangorestframework==3.15.2
psycopg2==2.9.10
django==5.1.7